Code cleanup
Remove dead code, commented out code
Clean up compile warnings

This is not a noop change as some changes to fix compile warnings actually change code paths. Hopefully for the better. Might even fix a few crash bugs due to incompatible pointer types being fixed.

+++ AirBitz/SlideoutView.m
@@ -304,7 +304,7 @@
 
 - (BOOL)gestureRecognizerShouldBegin:(UIGestureRecognizer *)gestureRecognizer {
     CGPoint translation = [(UIPanGestureRecognizer *)gestureRecognizer translationInView:gestureRecognizer.view.superview];
-    return fabsf(translation.x) > fabsf(translation.y);
+    return fabs(translation.x) > fabs(translation.y);
 }
 
 - (void)handleRecognizer:(UIPanGestureRecognizer *)recognizer fromBlock:(bool) block
